requirements.txt
streamlit==1.48.1
pypdf2==3.0.1
langchain==0.3.27
chromadb==0.5.24
sentence-transformers==3.6.0
python-docx==1.1.2
textblob==0.18.0